Testing infra for metalkubed
This is the basic Protractor (Selenium) infrastructure we use to test
the baremetal features in the web-ui fork. It is dependent on some
kubevirt code which is shared among the infrastructures, so those deps
are part of this PR as well.

The actual test is in tests/metalkube/dashboard.scenario.ts and it only
reads some of the counters in the dashboards page, and compares them to
what we get from the CLI.

This PR includes PRs 2287 and 2288 from Dan Trainor, who added custom
data-test-id attributes on the elements inthe dashboards, to support
the automation so it will be easier to find the elements.

This is still WIP because the dashboards are still not finalized. For
example the BM hosts counters in the inventory card are in "other"
state instead of OK state.

Basic steps to run the tests:

yarn install
yarn webdriver-update
export BRIDGE_BASE_ADDRESS='https://*.*.*.redhat.com'
export BRIDGE_AUTH_USERNAME=kubeadmin
export BRIDGE_AUTH_PASSWORD=22Z....
oc login -s *.*.*.redhat.com:6443 -u kubeadmin -p 22ZW....
export KUBECONFIG=/home/ukalifon/.kube/config
export NO_HEADLESS=true    # optional - if you want to see the browser
yarn run test-suite --suite baremetalSmokeTests --params.openshift true